refactor : edit Lotto package structure
